+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityTableModelTest.java
@@ -146,10 +146,10 @@ public final class SwingEntityTableModelTest extends AbstractEntityTableModelTes
 						.with(Department.ID, 1)
 						.with(Department.NAME, "dept")
 						.build();
-		assertThrows(IllegalArgumentException.class, () -> tableModel.items().addItems(singletonList(dept)));
+		assertThrows(IllegalArgumentException.class, () -> tableModel.items().add(singletonList(dept)));
 		assertThrows(IllegalArgumentException.class, () -> tableModel.items().visible().addItemsAt(0, singletonList(dept)));
 
-		assertThrows(NullPointerException.class, () -> tableModel.items().addItems(singletonList(null)));
+		assertThrows(NullPointerException.class, () -> tableModel.items().add(singletonList(null)));
 		assertThrows(NullPointerException.class, () -> tableModel.items().visible().addItemsAt(0, singletonList(null)));
 	}